+++ crypto/keymanager.go
