+++ crypto/keymanager_test.go
